# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Repair Story Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserResponseRole =
  (typeof UserResponseRole)[keyof typeof UserResponseRole];

export const UserResponseRole = {
  client: "client",
  technician: "technician",
  admin: "admin",
} as const;

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  role: UserResponseRole;
}

export interface AuthResponse {
  accessToken: string;
  user: UserResponse;
}

export interface CreateBookingBody {
  device: string;
  issue: string;
  appointmentAt?: string | null;
}

export type BookingResponseStatus =
  (typeof BookingResponseStatus)[keyof typeof BookingResponseStatus];

export const BookingResponseStatus = {
  new: "new",
  accepted: "accepted",
  diagnosing: "diagnosing",
  repairing: "repairing",
  ready: "ready",
  closed: "closed",
} as const;

export interface BookingResponse {
  id: number;
  code: string;
  device: string;
  issue: string;
  status: BookingResponseStatus;
  appointmentAt?: string | null;
  clientId: number;
  technicianId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface RepairLogResponse {
  id: number;
  bookingId: number;
  fromStatus: string;
  toStatus: string;
  note?: string | null;
  byUserId: number;
  byUserName: string;
  createdAt: string;
}

export interface BookingWithLogsResponse {
  booking: BookingResponse;
  logs: RepairLogResponse[];
}

export type UpdateStatusBodyTo =
  (typeof UpdateStatusBodyTo)[keyof typeof UpdateStatusBodyTo];

export const UpdateStatusBodyTo = {
  accepted: "accepted",
  diagnosing: "diagnosing",
  repairing: "repairing",
  ready: "ready",
  closed: "closed",
} as const;

export interface UpdateStatusBody {
  to: UpdateStatusBodyTo;
  note?: string;
}

export interface AssignTechBody {
  technicianId: number;
}

export interface CalendarSlot {
  time: string;
  available: boolean;
}

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  technicianId?: number;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  new: "new",
  accepted: "accepted",
  diagnosing: "diagnosing",
  repairing: "repairing",
  ready: "ready",
  closed: "closed",
} as const;

export type GetCalendarSlotsParams = {
  date: string;
};