# lib/api-zod/src/generated/types/bookingWithLogsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Repair Story Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingResponse } from "./bookingResponse";
import type { RepairLogResponse } from "./repairLogResponse";

export interface BookingWithLogsResponse {
  booking: BookingResponse;
  logs: RepairLogResponse[];
}