# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Repair Story Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  device: string;
  issue: string;
  appointmentAt?: Date | null;
}