# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Repair Story Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserResponseRole } from "./userResponseRole";

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  role: UserResponseRole;
}